backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
websockets
sqlalchemy
psycopg2-binary
httpx